Refactor server code and update dependencies
