+++ server/package.json
@@ -1,17 +1,18 @@
 {
   "name": "amazu-server",
   "version": "1.0.0",
-  "main": "dist/index.js",
+  "main": "dist/server.js",
   "license": "MIT",
   "scripts": {
     "build": "rimraf dist && tsc",
-    "start": "node dist/index.js",
-    "serve": "tsc-watch --onSuccess \"nodemon dist/index.js\"",
+    "start": "tsc-watch --onSuccess \"nodemon dist/server.js\"",
+    "serve": "nodemon ./src/server.ts",
     "clean": "rimraf dist",
     "lint": "eslint src/**/*.ts",
     "lint:fix": "eslint src/**/*.ts --fix"
   },
   "devDependencies": {
+    "@swc/core": "^1.3.101",
     "@types/cors": "^2.8.14",
     "@types/express": "^4.17.17",
     "@types/morgan": "^1.9.5",
@@ -26,6 +27,7 @@
     "morgan": "^1.10.0",
     "nodemon": "^3.0.1",
     "rimraf": "^5.0.1",
+    "ts-node": "^10.9.2",
     "tsc-watch": "^6.0.4",
     "typescript": "^5.2.2"
   },
