+++ server/swagger.ts
@@ -1,4 +1,3 @@
-import swaggerUi from 'swagger-ui-express';
 import swaggerDocs from './swagger.json';
 
-export { swaggerUi, swaggerDocs };
+export { swaggerDocs };
